requirements.txt
click>=8.0.0
flask>=2.0.0
jinja2>=3.0.0
matplotlib>=3.4.0
networkx>=2.5.0
numpy>=1.20.0
pillow>=8.2.0
pyparsing>=2.4.7
six>=1.16.0
sympy>=1.8.0
werkzeug>=2.0.0
cupy-cuda12x>=12.0.0  # Use the appropriate CUDA version for your RTX 3060
